Refactoring folder placement of components
